+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Make image uploading work
+
 ## 0.0.8 (2015-04-28)
 - Fetch VariantLocation#committed_stock as committed